cmd/snapd/tool/snap-recovery-chooser: move the source around in preparation for merge
