+++ cmd/snapd/tool/snap-recovery-chooser/export_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package main
+package snap_recovery_chooser
 
 import (
 	"io"
